requirements.txt
# Requirements for VRehab - Mind-Controlled Robot System
# Dependencias para VRehab - Sistema de Robot Controlado por la Mente

# Core data processing libraries
# Bibliotecas principales de procesamiento de datos
numpy>=1.21.0
pandas>=1.3.0

# Machine learning libraries
# Bibliotecas de machine learning
scikit-learn>=1.0.0

# EEG data streaming (Lab Streaming Layer)
# Streaming de datos EEG (Lab Streaming Layer)
pylsl>=1.16.0

# Robot control (Dobot Magic)
# Control de robot (Dobot Magic)
pydobot>=1.0.0

# Serial communication
# Comunicación serial
pyserial>=3.5

# GUI libraries
# Bibliotecas de interfaz gráfica
matplotlib>=3.5.0

# Data persistence
# Persistencia de datos
pickle-mixin>=1.0.2
